refactor: PaymentSuccessScreen을 event/admission으로 이동 (#509)
event_application_wizard_page.dart에서 payment feature의
PaymentSuccessScreen을 직접 import하는 cross-feature 의존성 해소.

PaymentSuccessScreen은 이벤트 신청 결제 완료 결과를 표시하는 UI로
도메인 상 event/admission feature에 속하므로 이동.

Before: apps/app_user/lib/src/features/payment/ui/payment_success_screen.dart
After:  apps/app_user/lib/src/features/event/admission/payment_success_screen.dart

Closes #509

Constraint: payment/ui/ 아래 해당 파일을 참조하는 곳이 wizard_page 1건뿐임을 확인
Rejected: PaymentCoordinator에 navigation 위임 | coordinator가 PaymentSuccessScreen을 import하여 동일 문제 반복
Rejected: GoRouter route로 전환 | Event 객체 전달 복잡도 증가, over-engineering
Confidence: high
Scope-risk: narrow
